admin server startup moved to admin.py
